+++ crates/maelstrom-test-runner/src/ui/simple/progress.rs
@@ -1,13 +1,9 @@
 mod multiple_progress_bars;
 mod no_bar;
-mod quiet_no_bar;
-mod quiet_progress_bar;
 mod test_listing;
 
 pub use multiple_progress_bars::MultipleProgressBars;
 pub use no_bar::NoBar;
-pub use quiet_no_bar::QuietNoBar;
-pub use quiet_progress_bar::QuietProgressBar;
 pub use test_listing::{TestListingProgress, TestListingProgressNoSpinner};
 
 use super::Terminal;