+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (kebab-case)
 - UI now has improved tracking of the state of tests. The fancy UI in particular will now only show
   a test as running once it has actually started executing.
+- `quiet` configuration option removed from test runners, replaced with new "quiet" UI.
 
 ### Fixed
 - Fixed issue with test-runner fancy UI where it would crash if the summary was too large.